progress on trending dapps
